Refactoring the packages into different folders structure
